irctc_backend/requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.5
Django==5.2.11
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
mysqlclient==2.2.8
pycparser==3.0
PyJWT==2.11.0
pymongo==4.16.0
PyMySQL==1.1.2
python-dotenv==1.2.1
sqlparse==0.5.5
